Clean TQueue
