+++ src/nrniv/savstate.cpp
@@ -10,7 +10,7 @@
 #include "ndatclas.h"
 #include "nrniv_mf.h"
 
-#include "tqueue.h"
+#include "tqueue.hpp"
 #include "netcon.h"
 #include "vrecitem.h"
 #include "utils/enumerate.h"
